fix: deep audit round 4 — data races, validation gaps, and defensive hardening
- Fix data race in DiffLayer parent re-parenting: hold DiffLayer.lock during
  flattenOldest re-parent to prevent concurrent Account/Storage reads from
  seeing stale parent pointer
- Fix DiffLayer Account/Storage: capture parent reference under lock before
  delegating to parent layer
- Fix snapshot persistDiffToDisk: abort transaction on proto.Marshal failure
  instead of silently continuing with incomplete data
- Fix snapshot generator: only mark genReady after verifying generation
  actually completed (not cancelled or crashed)
- Fix Pacemaker latencyCursor: use uint64 to prevent signed integer overflow
  producing negative ring buffer index
- Fix processCommitVote: add voter index validation against validator set,
  matching processVote's defensive pattern
- Fix handleFutureViewTimeout: call Timeout() before advanceToView so
  buffered message replay sees correct phase and backoff level
- Fix PeerDAS service: protect lastSampledBlock with atomic.Uint64 and
  blockProvider with mutex to prevent data races in sampling loop

+++ internal/consensus/hotstuff/pacemaker.go
@@ -25,7 +25,7 @@ type Pacemaker struct {
 
 	// Adaptive timeout: ring buffer of recent commit latencies.
 	recentLatencies []time.Duration
-	latencyCursor   int
+	latencyCursor   uint64
 }
 
 const (
@@ -59,12 +59,13 @@ func (p *Pacemaker) ObserveCommitLatency(latency time.Duration) {
 }
 
 // effectiveBaseTimeout returns the adaptive base timeout considering observed latencies.
+// Caller must hold at least p.mu.RLock.
 func (p *Pacemaker) effectiveBaseTimeout() time.Duration {
 	if len(p.recentLatencies) < 5 {
 		return p.baseTimeout
 	}
 
-	// Compute p95 of recent latencies.
+	// Compute p95 of recent latencies. Copy the slice to avoid sorting the original.
 	sorted := make([]int64, len(p.recentLatencies))
 	for i, d := range p.recentLatencies {
 		sorted[i] = d.Milliseconds()
